refactor(api): move target profile summary toute and controller to prescription
