+++ api/tests/prescription/target-profile/unit/application/admin-target-profile-controller_test.js
@@ -268,4 +268,41 @@ describe('Unit | Controller | admin-target-profile-controller', function () {
       });
     });
   });
+
+  describe('#findTargetProfileSummariesForAdmin', function () {
+    it('should return serialized summaries', async function () {
+      // given
+      sinon.stub(prescriptionTargetProfileUsecases, 'findOrganizationTargetProfileSummariesForAdmin');
+      const request = {
+        params: { id: 123 },
+      };
+      const targetProfileSummary = domainBuilder.buildTargetProfileSummaryForAdmin({
+        id: 456,
+        name: 'Super profil cible',
+        outdated: false,
+      });
+      prescriptionTargetProfileUsecases.findOrganizationTargetProfileSummariesForAdmin
+        .withArgs({ organizationId: 123 })
+        .resolves([targetProfileSummary]);
+
+      // when
+      const response = await targetProfileController.findTargetProfileSummariesForAdmin(request);
+
+      // then
+      expect(response).to.deep.equal({
+        data: [
+          {
+            type: 'target-profile-summaries',
+            id: '456',
+            attributes: {
+              name: 'Super profil cible',
+              outdated: false,
+              'can-detach': false,
+              'created-at': undefined,
+            },
+          },
+        ],
+      });
+    });
+  });
 });
