+++ api/tests/unit/application/organizations/index_test.js
@@ -137,109 +137,4 @@ describe('Unit | Router | organization-router', function () {
       sinon.assert.notCalled(organizationController.archiveOrganization);
     });
   });
-
-  describe('GET /api/admin/organizations/{id}/target-profile-summaries', function () {
-    it('should allow to controller if user has role SUPER_ADMIN', async function () {
-      // given
-      sinon.stub(organizationController, 'findTargetProfileSummariesForAdmin').returns('ok');
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleSupport')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleMetier')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon.stub(securityPreHandlers, 'checkAdminMemberHasRoleSuperAdmin').callsFake((request, h) => h.response(true));
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      await httpTestServer.request('GET', '/api/admin/organizations/1/target-profile-summaries');
-
-      // then
-      sinon.assert.calledOnce(organizationController.findTargetProfileSummariesForAdmin);
-    });
-
-    it('should allow to controller if user has role SUPPORT', async function () {
-      // given
-      sinon.stub(organizationController, 'findTargetProfileSummariesForAdmin').returns('ok');
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleSupport')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleMetier')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon.stub(securityPreHandlers, 'checkAdminMemberHasRoleSuperAdmin').callsFake((request, h) => h.response(true));
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      await httpTestServer.request('GET', '/api/admin/organizations/1/target-profile-summaries');
-
-      // then
-      sinon.assert.calledOnce(organizationController.findTargetProfileSummariesForAdmin);
-    });
-
-    it('should allow to controller if user has role METIER', async function () {
-      // given
-      sinon.stub(organizationController, 'findTargetProfileSummariesForAdmin').returns('ok');
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleSuperAdmin')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleSupport')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon.stub(securityPreHandlers, 'checkAdminMemberHasRoleMetier').callsFake((request, h) => h.response(true));
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      await httpTestServer.request('GET', '/api/admin/organizations/1/target-profile-summaries');
-
-      // then
-      sinon.assert.calledOnce(organizationController.findTargetProfileSummariesForAdmin);
-    });
-
-    it('should return 403 without reaching controller if user has not an allowed role', async function () {
-      // given
-      sinon.stub(organizationController, 'findTargetProfileSummariesForAdmin').returns('ok');
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleSuperAdmin')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleSupport')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleMetier')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const response = await httpTestServer.request('GET', '/api/admin/organizations/1/target-profile-summaries');
-
-      // then
-      expect(response.statusCode).to.equal(403);
-      sinon.assert.notCalled(organizationController.findTargetProfileSummariesForAdmin);
-    });
-
-    it('should return a 400 HTTP response when organization id is not valid', async function () {
-      // given
-      sinon.stub(organizationController, 'findTargetProfileSummariesForAdmin').returns('ok');
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleSuperAdmin')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon
-        .stub(securityPreHandlers, 'checkAdminMemberHasRoleSupport')
-        .callsFake((request, h) => h.response({ errors: new Error('forbidden') }).code(403));
-      sinon.stub(securityPreHandlers, 'checkAdminMemberHasRoleMetier').callsFake((request, h) => h.response(true));
-      const httpTestServer = new HttpTestServer();
-      await httpTestServer.register(moduleUnderTest);
-
-      // when
-      const response = await httpTestServer.request('GET', '/api/admin/organizations/coucou/target-profile-summaries');
-
-      // then
-      expect(response.statusCode).to.equal(400);
-    });
-  });
 });
